+++ services/common/src/collections/index.ts
@@ -3,6 +3,7 @@ import mongoose from 'mongoose';
 export { default as Account } from './account.js';
 export { default as Document } from './document.js';
 export { default as Email } from './email.js';
+export { default as Expense } from './expense.js';
 export { default as Lease } from './lease.js';
 export { default as Property } from './property.js';
 export { default as Realm } from './realm.js';